+++ daabbcc3d/src/daabbcc3d/daabbcc.cpp
@@ -94,7 +94,7 @@ namespace daabbcc3d
     // Proxy Operations
     ////////////////////////////////////////
 
-    int32_t AddProxy(uint8_t groupID, float x, float y, float z, uint32_t width, uint32_t height, uint32_t depth, uint64_t categoryBits)
+    int32_t AddProxy(uint8_t groupID, float x, float y, float z, float width, float height, float depth, uint64_t categoryBits)
     {
         Bound(&m_daabbcc.m_aabb, x, y, z, width, height, depth);
         int32_t proxyID = b2DynamicTree_CreateProxy(&m_daabbcc.m_treeGroup->m_dynamicTree, m_daabbcc.m_aabb, categoryBits, groupID);
@@ -102,7 +102,7 @@ namespace daabbcc3d
         return proxyID;
     }
 
-    void AddGameObject(uint8_t groupID, int32_t proxyID, dmVMath::Point3 position, uint32_t width, uint32_t height, uint32_t depth, dmGameObject::HInstance gameObjectInstance, bool getWorldPosition)
+    void AddGameObject(uint8_t groupID, int32_t proxyID, dmVMath::Point3 position, float width, float height, float depth, dmGameObject::HInstance gameObjectInstance, bool getWorldPosition)
     {
         GameObject gameObject;
 
@@ -125,13 +125,13 @@ namespace daabbcc3d
         }
     }
 
-    void MoveProxy(int32_t proxyID, float x, float y, float z, uint32_t width, uint32_t height, uint32_t depth)
+    void MoveProxy(int32_t proxyID, float x, float y, float z, float width, float height, float depth)
     {
         Bound(&m_daabbcc.m_aabb, x, y, z, width, height, depth);
         b2DynamicTree_MoveProxy(&m_daabbcc.m_treeGroup->m_dynamicTree, proxyID, m_daabbcc.m_aabb);
     }
 
-    void UpdateGameobjectSize(uint8_t groupID, int32_t proxyID, uint32_t width, uint32_t height, uint32_t depth)
+    void UpdateGameobjectSize(uint8_t groupID, int32_t proxyID, float width, float height, float depth)
     {
         for (uint32_t i = 0; i < m_daabbcc.m_gameObjectContainer.Size(); ++i)
         {
@@ -260,7 +260,7 @@ namespace daabbcc3d
     // QUERIES
     /**************************/
 
-    void QueryAABB(float x, float y, float z, uint32_t width, uint32_t height, uint32_t depth, uint64_t maskBits, bool isManifold)
+    void QueryAABB(float x, float y, float z, float width, float height, float depth, uint64_t maskBits, bool isManifold)
     {
         Bound(&m_daabbcc.m_aabb, x, y, z, width, height, depth);
 
@@ -278,7 +278,7 @@ namespace daabbcc3d
         Query(&m_daabbcc.m_aabb, QueryCallback, &m_daabbcc.m_queryContainer, maskBits);
     }
 
-    void QueryAABBSort(float x, float y, float z, uint32_t width, uint32_t height, uint32_t depth, uint64_t maskBits, bool isManifold)
+    void QueryAABBSort(float x, float y, float z, float width, float height, float depth, uint64_t maskBits, bool isManifold)
     {
         Bound(&m_daabbcc.m_aabb, x, y, z, width, height, depth);
 
@@ -624,7 +624,7 @@ namespace daabbcc3d
         m->n = n;
     }
 
-    static inline void Bound(b2AABB* aabb, float x, float y, float z, uint32_t width, uint32_t height, uint32_t depth)
+    static inline void Bound(b2AABB* aabb, float x, float y, float z, float width, float height, float depth)
     {
         aabb->lowerBound = { x - (width / 2.0f), y - (height / 2.0f), z - (depth / 2.0f) };
         aabb->upperBound = { x + (width / 2.0f), y + (height / 2.0f), z + (depth / 2.0f) };
